Add x-option to eager render tabs
By default, v-tabs are lazy loaded, which prevents the validation to
happen in tabs that weren't rendered yet.

If you want all validations to happen at the initial rendering of the
form, you must eager render all tabs.

This impacts the rendering performance and that's why it will continue
as lazy by default.