feat: Add option to render descriptions as links
Sometimes, help descriptions are too long for a small tooltip and
are better of at a different location. In these cases it would be
nice to provide a direct link (opening in a new tab) instead.

The option "linkTooltipsIfAvailable" does exactly that.
The old behaviour is preserved if the option is false (default).